requirements.txt
Flask==2.0.1
pygame==2.1.0
Werkzeug==2.0.3
python-dotenv==1.0.0
sqlalchemy==1.4.31
gunicorn==22.0.0